[Refactor] Move Network to core/graph
